Move log_medication_taken and undo_medication_taken from medication.py to administration_log.py
